requirements.txt
Django==1.6.5
django-tastypie==0.11.0
ipython==2.0.0
django-timedeltafield